refactor(brett): extract figures.ts (figureMaps, applyMutation, locks)
